+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Implement `;!` to execute a shell command from the command bar
 * Disabled continuous deployment in GHA. See `.github/workflows/build_artifact.yml`
 for details
+* `M<char>` now marks the current row to the given letter on the keyboard
 
 * Resolve crash when user tries to switch top or bottom line outside of scope
 * Resolved crash when the user tries to `delete word` past the end of a line